Added acceleration capping to the Ballio elevator
In case the pulleys slip and it stalls. The non-elevator and non-drive
motors are insignificant enough not to need acceleration capping (at
least, until channel current-dependent acceleration capping is merged).
This should be added to the shooter’s indexer as well in the respective
shooter branch, and a discussion about climber acceleration capping is
pending (and that change is outside the scope of this branch anyway).